restructure client mod
