+++ .pre-commit-config.yaml
@@ -15,4 +15,4 @@ repos:
     hooks:
       - id: fmt
       - id: cargo-check
-      # - id: clippy
+      - id: clippy
