+++ core/src/client/mod.rs
@@ -1,13 +1,10 @@
-//! Generic implementation of ChannelBuilder and SparkConnectClient
-
-use std::collections::HashMap;
-use std::env;
-use std::str::FromStr;
-use std::sync::Arc;
-
 use crate::errors::SparkError;
+
 use crate::spark;
 use spark::execute_plan_response::ResponseType;
+use std::sync::Arc;
+use tokio::sync::RwLock;
+use tonic::codegen::{Body, Bytes, StdError};
 
 use spark::spark_connect_service_client::SparkConnectServiceClient;
 
@@ -16,273 +13,11 @@ use arrow::error::ArrowError;
 use arrow::record_batch::RecordBatch;
 use arrow_ipc::reader::StreamReader;
 
-use tonic::codegen::{Body, Bytes, StdError};
-use tonic::metadata::{
-    Ascii, AsciiMetadataValue, KeyAndValueRef, MetadataKey, MetadataMap, MetadataValue,
-};
-use tonic::service::Interceptor;
-use tonic::Status;
+mod builder;
+mod middleware;
 
-use tokio::sync::RwLock;
-
-use url::Url;
-
-use uuid::Uuid;
-
-type Host = String;
-type Port = u16;
-type UrlParse = (Host, Port, Option<HashMap<String, String>>);
-
-/// ChannelBuilder validates a connection string
-/// based on the requirements from [Spark Documentation](https://github.com/apache/spark/blob/master/connector/connect/docs/client-connection-string.md)
-#[derive(Clone, Debug)]
-pub struct ChannelBuilder {
-    host: Host,
-    port: Port,
-    session_id: Uuid,
-    token: Option<String>,
-    user_id: Option<String>,
-    user_agent: Option<String>,
-    use_ssl: bool,
-    headers: Option<MetadataMap>,
-}
-
-impl Default for ChannelBuilder {
-    fn default() -> Self {
-        let connection = match env::var("SPARK_REMOTE") {
-            Ok(conn) => conn.to_string(),
-            Err(_) => "sc://localhost:15002".to_string(),
-        };
-
-        ChannelBuilder::create(&connection).unwrap()
-    }
-}
-
-impl ChannelBuilder {
-    pub fn new() -> Self {
-        ChannelBuilder::default()
-    }
-
-    pub fn endpoint(&self) -> String {
-        let scheme = if cfg!(feature = "tls") {
-            "https"
-        } else {
-            "http"
-        };
-
-        format!("{}://{}:{}", scheme, self.host, self.port)
-    }
-
-    pub fn token(&self) -> Option<String> {
-        self.token.to_owned()
-    }
-
-    pub fn headers(&self) -> Option<MetadataMap> {
-        self.headers.to_owned()
-    }
-
-    fn create_user_agent(user_agent: Option<&str>) -> Option<String> {
-        let user_agent = user_agent.unwrap_or("_SPARK_CONNECT_RUST");
-        let pkg_version = env!("CARGO_PKG_VERSION");
-        let os = env::consts::OS.to_lowercase();
-
-        Some(format!(
-            "{} os/{} spark_connect_rs/{}",
-            user_agent, os, pkg_version
-        ))
-    }
-
-    fn create_user_id(user_id: Option<&str>) -> Option<String> {
-        match user_id {
-            Some(user_id) => Some(user_id.to_string()),
-            None => match env::var("USER") {
-                Ok(user) => Some(user),
-                Err(_) => None,
-            },
-        }
-    }
-
-    pub fn parse_connection_string(connection: &str) -> Result<UrlParse, SparkError> {
-        let url = Url::parse(connection).map_err(|_| {
-            SparkError::InvalidConnectionUrl("Failed to parse the connection URL".to_string())
-        })?;
-
-        if url.scheme() != "sc" {
-            return Err(SparkError::InvalidConnectionUrl(
-                "The URL must start with 'sc://'. Please update the URL to follow the correct format, e.g., 'sc://hostname:port'".to_string(),
-            ));
-        };
-
-        let host = url
-            .host_str()
-            .ok_or_else(|| {
-                SparkError::InvalidConnectionUrl(
-                    "The hostname must not be empty. Please update
-                    the URL to follow the correct format, e.g., 'sc://hostname:port'."
-                        .to_string(),
-                )
-            })?
-            .to_string();
-
-        let port = url.port().ok_or_else(|| {
-            SparkError::InvalidConnectionUrl(
-                "The port must not be empty. Please update
-                    the URL to follow the correct format, e.g., 'sc://hostname:port'."
-                    .to_string(),
-            )
-        })?;
-
-        let headers = ChannelBuilder::parse_headers(url);
-
-        Ok((host, port, headers))
-    }
-
-    pub fn parse_headers(url: Url) -> Option<HashMap<String, String>> {
-        let path: Vec<&str> = url
-            .path()
-            .split(';')
-            .filter(|&pair| (pair != "/") & (!pair.is_empty()))
-            .collect();
-
-        if path.is_empty() || (path.len() == 1 && (path[0].is_empty() || path[0] == "/")) {
-            return None;
-        }
-
-        let headers: HashMap<String, String> = path
-            .iter()
-            .copied()
-            .map(|pair| {
-                let mut parts = pair.splitn(2, '=');
-                (
-                    parts.next().unwrap_or("").to_string(),
-                    parts.next().unwrap_or("").to_string(),
-                )
-            })
-            .collect();
-
-        if headers.is_empty() {
-            return None;
-        }
-
-        Some(headers)
-    }
-
-    /// Create and validate a connnection string
-    #[allow(unreachable_code)]
-    pub fn create(connection: &str) -> Result<ChannelBuilder, SparkError> {
-        let (host, port, headers) = ChannelBuilder::parse_connection_string(connection)?;
-
-        let mut channel_builder = ChannelBuilder {
-            host,
-            port,
-            session_id: Uuid::new_v4(),
-            token: None,
-            user_id: ChannelBuilder::create_user_id(None),
-            user_agent: ChannelBuilder::create_user_agent(None),
-            use_ssl: false,
-            headers: None,
-        };
-
-        let mut headers = match headers {
-            Some(headers) => headers,
-            None => return Ok(channel_builder),
-        };
-
-        channel_builder.user_id = headers
-            .remove("user_id")
-            .map(|user_id| ChannelBuilder::create_user_id(Some(&user_id)))
-            .unwrap_or_else(|| ChannelBuilder::create_user_id(None));
-
-        channel_builder.user_agent = headers
-            .remove("user_agent")
-            .map(|user_agent| ChannelBuilder::create_user_agent(Some(&user_agent)))
-            .unwrap_or_else(|| ChannelBuilder::create_user_agent(None));
-
-        if let Some(token) = headers.remove("token") {
-            channel_builder.token = Some(format!("Bearer {token}"));
-        }
-
-        if let Some(session_id) = headers.remove("session_id") {
-            channel_builder.session_id = Uuid::from_str(&session_id).unwrap()
-        }
-
-        if let Some(use_ssl) = headers.remove("use_ssl") {
-            if use_ssl.to_lowercase() == "true" {
-                #[cfg(not(feature = "tls"))]
-                {
-                    panic!(
-                        "The 'use_ssl' option requires the 'tls' feature, but it's not enabled!"
-                    );
-                };
-                channel_builder.use_ssl = true
-            }
-        };
-
-        channel_builder.headers = Some(metadata_builder(&headers));
-
-        Ok(channel_builder)
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct MetadataInterceptor {
-    token: Option<String>,
-    metadata: Option<MetadataMap>,
-}
-
-impl Interceptor for MetadataInterceptor {
-    fn call(&mut self, mut req: tonic::Request<()>) -> Result<tonic::Request<()>, Status> {
-        if let Some(header) = &self.metadata {
-            merge_metadata(req.metadata_mut(), header);
-        }
-        if let Some(token) = &self.token {
-            req.metadata_mut().insert(
-                "authorization",
-                AsciiMetadataValue::from_str(token.as_str()).unwrap(),
-            );
-        }
-
-        Ok(req)
-    }
-}
-
-impl MetadataInterceptor {
-    pub fn new(token: Option<String>, metadata: Option<MetadataMap>) -> Self {
-        MetadataInterceptor { token, metadata }
-    }
-}
-
-fn metadata_builder(headers: &HashMap<String, String>) -> MetadataMap {
-    let mut metadata_map = MetadataMap::new();
-    for (key, val) in headers.iter() {
-        let meta_val = MetadataValue::from_str(val.as_str()).unwrap();
-        let meta_key = MetadataKey::from_str(key.as_str()).unwrap();
-
-        metadata_map.insert(meta_key, meta_val);
-    }
-
-    metadata_map
-}
-
-fn merge_metadata(metadata_into: &mut MetadataMap, metadata_from: &MetadataMap) {
-    metadata_for_each(metadata_from, |key, value| {
-        if key.to_string().starts_with("x-") {
-            metadata_into.insert(key, value.to_owned());
-        }
-    })
-}
-
-fn metadata_for_each<F>(metadata: &MetadataMap, mut f: F)
-where
-    F: FnMut(&MetadataKey<Ascii>, &MetadataValue<Ascii>),
-{
-    for kv_ref in metadata.iter() {
-        match kv_ref {
-            KeyAndValueRef::Ascii(key, value) => f(key, value),
-            KeyAndValueRef::Binary(_key, _value) => {}
-        }
-    }
-}
+pub use builder::ChannelBuilder;
+pub use middleware::MetadataInterceptor;
 
 #[allow(dead_code)]
 #[derive(Default, Debug, Clone)]
@@ -319,45 +54,6 @@ pub struct AnalyzeHandler {
     pub(crate) get_storage_level: Option<spark::StorageLevel>,
 }
 
-impl ResponseHandler {
-    fn new() -> Self {
-        Self {
-            session_id: None,
-            operation_id: None,
-            response_id: None,
-            metrics: None,
-            observed_metrics: None,
-            schema: None,
-            batches: Vec::new(),
-            sql_command_result: None,
-            write_stream_operation_start_result: None,
-            streaming_query_command_result: None,
-            get_resources_command_result: None,
-            streaming_query_manager_command_result: None,
-            result_complete: false,
-            total_count: 0,
-        }
-    }
-}
-
-impl AnalyzeHandler {
-    fn new() -> Self {
-        Self {
-            schema: None,
-            explain: None,
-            tree_string: None,
-            is_local: None,
-            is_streaming: None,
-            input_files: None,
-            spark_version: None,
-            ddl_parse: None,
-            same_semantics: None,
-            semantic_hash: None,
-            get_storage_level: None,
-        }
-    }
-}
-
 #[derive(Clone, Debug)]
 pub struct SparkConnectClient<T> {
     stub: Arc<RwLock<SparkConnectServiceClient<T>>>,
@@ -382,8 +78,8 @@ where
         SparkConnectClient {
             stub,
             builder,
-            handler: ResponseHandler::new(),
-            analyzer: AnalyzeHandler::new(),
+            handler: ResponseHandler::default(),
+            analyzer: AnalyzeHandler::default(),
             user_context: Some(spark::UserContext {
                 user_id: user_ref.clone(),
                 user_name: user_ref,
@@ -451,7 +147,7 @@ where
         drop(client);
 
         // clear out any prior responses
-        self.handler = ResponseHandler::new();
+        self.handler = ResponseHandler::default();
 
         while let Some(_resp) = match stream.message().await {
             Ok(Some(msg)) => {
@@ -463,9 +159,7 @@ where
             Ok(None) => {
                 if self.use_reattachable_execute && !self.handler.result_complete {
                     println!("didn't get the whole stream submitting a reattach");
-                    if let Err(err) = self.reattach_request().await {
-                        return Err(err);
-                    }
+                    self.reattach_request().await?
                 }
                 None
             }
@@ -506,7 +200,7 @@ where
         req.analyze = Some(analyze);
 
         // clear out any prior responses
-        self.analyzer = AnalyzeHandler::new();
+        self.analyzer = AnalyzeHandler::default();
 
         let mut client = self.stub.write().await;
         let resp = client.analyze_plan(req).await?.into_inner();
@@ -565,9 +259,6 @@ where
         let mut stream = client.reattach_execute(req).await?.into_inner();
         drop(client);
 
-        // clear out any prior responses
-        // self.handler = ResponseHandler::new();
-
         while let Some(_resp) = match stream.message().await {
             Ok(Some(msg)) => {
                 // Handle the message
@@ -578,9 +269,7 @@ where
             Ok(None) => {
                 if self.use_reattachable_execute && !self.handler.result_complete {
                     println!("something bad happened in the reattach. submitting reattach");
-                    if let Err(err) = Box::pin(self.reattach_request()).await {
-                        return Err(err);
-                    }
+                    Box::pin(self.reattach_request()).await?
                 }
                 None
             }
@@ -892,50 +581,5 @@ where
         self.analyzer.get_storage_level.to_owned().ok_or_else(|| {
             SparkError::AnalysisException("Storage Level response is empty".to_string())
         })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_channel_builder_default() {
-        let expected_url = "http://localhost:15002".to_string();
-
-        let cb = ChannelBuilder::default();
-
-        assert_eq!(expected_url, cb.endpoint())
-    }
-
-    #[test]
-    fn test_panic_incorrect_url_scheme() {
-        let connection = "http://127.0.0.1:15002";
-
-        assert!(ChannelBuilder::create(connection).is_err())
-    }
-
-    #[test]
-    fn test_panic_missing_url_host() {
-        let connection = "sc://:15002";
-
-        assert!(ChannelBuilder::create(connection).is_err())
-    }
-
-    #[test]
-    fn test_panic_missing_url_port() {
-        let connection = "sc://127.0.0.1";
-
-        assert!(ChannelBuilder::create(connection).is_err())
-    }
-
-    #[test]
-    #[should_panic(
-        expected = "The 'use_ssl' option requires the 'tls' feature, but it's not enabled!"
-    )]
-    fn test_panic_ssl() {
-        let connection = "sc://127.0.0.1:443/;use_ssl=true";
-
-        ChannelBuilder::create(&connection).unwrap();
     }
 }
